Fixed IDs

+++ IRO.Task.NoteBase.BLL.Contracts/IBookLogic.cs
@@ -6,9 +6,9 @@ namespace IRO.Task.NoteBase.BLL.Contracts
     public interface IBookLogic
     {
         bool AddBook(Book book);
-        bool DeleteBook(int bookId);
+        bool DeleteBook(long bookId);
         List<Book> GetByUser(User user);
-        Book GetById(int bookId);
+        Book GetById(long bookId);
         List<Book> GetAll();
     }
 }
