+++ IRO.Task.NoteBase.BLL.Core/NoteLogic.cs
@@ -38,7 +38,7 @@ namespace IRO.Task.NoteBase.BLL.Core
             return true;
         }
 
-        public bool DeleteNote(int noteId)
+        public bool DeleteNote(long noteId)
         {
             var record = GetById(noteId);
             if (record == null)
@@ -51,6 +51,6 @@ namespace IRO.Task.NoteBase.BLL.Core
 
         public List<Note> GetAll() => _dbSet.ToList();
         public List<Note> GetByBook(Book book) => _dbSet.Where(x => x.ParentBookId == book.Id).ToList();
-        public Note GetById(int noteId) => _dbSet.FirstOrDefault(x => x.Id == noteId);
+        public Note GetById(long noteId) => _dbSet.FirstOrDefault(x => x.Id == noteId);
     }
 }
